requirements.txt
transformers==4.3.3
pytorch-lightning==1.2.6
pandas==1.2.3
numpy==1.18.5
nlp==0.4.0
sentencepiece==0.1.95
wandb
nltk